Add caching for imported assets
This will cache the imported files between each run. This massively speeds up the export stage for big projects.

Note that this will save the cache *after* a stage is done, so if you add new files it will still import them once for each platform (but will not import them again in the future, which is what matter).