Add missing step to C#/Mono tutorial in README.md
Previously, the detail that you had to change .stable to .stable.mono in the Setup steps directory structure was not listed and was very confusing for me until I found an issue from some time ago ( https://github.com/abarichello/godot-ci/issues/29#issuecomment-680315417 ) that stated this solution but was never addressed. I hope this can save people some trouble in the future.